+++ src/main/java/org/example/richardwebsite/service/OrderService.java
@@ -16,29 +16,30 @@ public class OrderService {
     }
 
     public Order createOrder(User user, Cart cart) {
+
         Integer max = orderRepository.findMaxUserOrderNumber(user.getId());
-        // Branches 1 & 2: max is null vs max has value
         int nextNumber = (max == null) ? 1 : max + 1;
 
         Order order = new Order();
         order.setUser(user);
         order.setCustomerName(user.getUsername());
-        order.setUserOrderNumber(nextNumber);
+
+        order.setUserOrderNumber(nextNumber); // FIXED (no String)
+
         order.setStatus(OrderStatus.PENDING);
 
-        // calculateTotal handles null/empty carts safely
         BigDecimal total = calculateTotal(cart);
         order.setTotal(total);
 
-        // FIX: Added null check to prevent NPE (Branches 5 & 6)
         if (cart != null && cart.getItems() != null) {
-            // Branches 7 & 8: Loop logic (has items vs no items)
             for (CartItem cartItem : cart.getItems()) {
+
                 OrderItem item = new OrderItem(
                         cartItem.getBook().getTitle(),
                         cartItem.getBook().getPrice(),
                         cartItem.getQuantity()
                 );
+
                 order.addItem(item);
             }
         }
@@ -55,7 +56,7 @@ public class OrderService {
 
         // Branches 9 & 10: Loop logic (has items vs no items)
         for (CartItem cartItem : cart.getItems()) {
-            BigDecimal price = BigDecimal.valueOf(cartItem.getBook().getPrice());
+            BigDecimal price = cartItem.getBook().getPrice();
             BigDecimal quantity = BigDecimal.valueOf(cartItem.getQuantity());
             total = total.add(price.multiply(quantity));
         }
